move activity recatngles to seperate files
closes #274
